Debugging GPU P2P kernel: Step 1.

+++ include/types.h
@@ -18,10 +18,10 @@
 #include "vec.h"
 
 typedef long                 bigint;                            //!< Big integer type
-//typedef float                real;                              //!< Real number type on CPU
-typedef double               real;                              //!< Real number type on CPU
-//typedef float                gpureal;                           //!< Real number type on GPU
-typedef double               gpureal;                           //!< Real number type on GPU
+typedef float                real;                              //!< Real number type on CPU
+//typedef double               real;                              //!< Real number type on CPU
+typedef float                gpureal;                           //!< Real number type on GPU
+//typedef double               gpureal;                           //!< Real number type on GPU
 typedef std::complex<double> complex;                           //!< Complex number type
 
 #ifndef KERNEL
